create helper classes to clean up adapter
