+++ app/src/main/java/com/example/chatapp/presentation/ui/chat/ChatFragment.kt
@@ -8,6 +8,8 @@ import com.example.chatapp.data.service.broadcast.BroadcastService
 import com.example.chatapp.databinding.FragmentChatBinding
 import com.example.chatapp.presentation.base.BaseFragment
 import com.example.chatapp.presentation.ui.chat.adapter.ChatAdapter
+import com.example.chatapp.presentation.ui.chat.adapter.util.RecyclerViewUtilClass
+import com.example.chatapp.presentation.ui.chat.adapter.util.RecyclerViewUtilClassImpl
 import com.example.chatapp.presentation.ui.chat.chat_user.ChatUser
 import com.example.chatapp.presentation.ui.chat.viewmodel.ChatViewModel
 import com.example.chatapp.util.Constants.MESSAGE_SENDER_KEY
@@ -23,8 +25,8 @@ class ChatFragment(private val user: ChatUser) :
 
     override fun getViewModelClass(): KClass<ChatViewModel> = ChatViewModel::class
 
-    private val messageAdapter by lazy { ChatAdapter(user.name) }
-
+    private val util: RecyclerViewUtilClass by lazy { RecyclerViewUtilClassImpl() }
+    private val messageAdapter by lazy { ChatAdapter(user.name, util) }
     override val broadcastService: BroadcastService = BroadcastService()
 
     override fun onBindViewModel(viewModel: ChatViewModel) {
